Move http server to separate file.
